fix(productListViewed): map productIds not product category to content_ids param
Previously we were mapping the product list category as the only member of an array to the FB pixel
content_ids param. FB expects this to be an array of productIds instead. This PR adds a change to
first check if the user has sent through a products array with productIds and if so, properly maps
them to the content_ids array. Otherwise, it simply falls back on the legacy behavior.

https://segment.atlassian.net/browse/PLATFORM-2400